Added latest and manul sidebars. Added category sidebar to sections
