+++ ubyssey/static_src/src/styles/components/_infinitefeed.scss
@@ -26,7 +26,7 @@
 .c-sidebar_container {
     height: 1500px;
     min-height: fit-content;
-    margin-top: 65px;
+    margin-bottom: 65px;
 }
 
 .c-sidebar {
